+++ wp-dapp.php
@@ -57,12 +57,6 @@ function wpdapp_version_update() {
         // Get current options
         $options = get_option('wpdapp_options', []);
         
-        // Ensure auto_publish is explicitly set to 0 by default
-        if (!isset($options['auto_publish'])) {
-            $options['auto_publish'] = 0;
-            update_option('wpdapp_options', $options);
-        }
-        
         // Update stored version
         update_option('wpdapp_version', WPDAPP_VERSION);
     }
@@ -73,11 +67,6 @@ add_action('plugins_loaded', 'wpdapp_version_update', 5); // Priority 5 to run b
  * Initialize plugin classes on plugins_loaded
  */
 function wpdapp_init() {
-    // Reset auto_publish option to ensure it's off by default
-    $options = get_option('wpdapp_options', []);
-    $options['auto_publish'] = 0;
-    update_option('wpdapp_options', $options);
-    
     // Initialize classes
     new WP_Dapp_Settings_Page();
     new WP_Dapp_Post_Meta();
@@ -102,7 +91,6 @@ function wpdapp_activate() {
         'default_beneficiary_account' => 'diggndeeper.com',
         'default_beneficiary_weight' => '100', // 1%
         'default_tags' => 'blog,wordpress',
-        'auto_publish' => '0'
     ];
     
     // Only set options if they don't exist